requirements.txt
PyYAML==6.0.2
streamlit==1.41.0
streamlit_authenticator==0.4.1
streamlit_option_menu==0.4.0
